+++ test/OpenTelemetry.AutoInstrumentation.Tests/Configurations/SettingsTests.cs
@@ -352,6 +352,7 @@ public class SettingsTests : IDisposable
 
     [Theory]
     [InlineData("CONTAINER", ResourceDetector.Container)]
+    [InlineData("AZUREAPPSERVICE", ResourceDetector.AzureAppService)]
     internal void GeneralSettings_Instrumentations_SupportedValues(string resourceDetector, ResourceDetector expectedResourceDetector)
     {
         Environment.SetEnvironmentVariable(ConfigurationKeys.ResourceDetectorEnabled, "false");